requirements.txt
google-genai
python-dotenv
playwright
Pillow
pydantic
validators
beautifulsoup4
asyncio
crawl4ai